Correções de bug e Validações.
Correções de alguns bugs e integração com Clipboard.

+++ Comos Utilities/classes/Arquivo.cs
@@ -9,31 +9,32 @@ namespace Comos_Utilities.classes
 {
     class Arquivo
     {
-
+        //Objeto do tipo FolderBrowserDialog para poder abrir a janela de seleção de diretório
         private FolderBrowserDialog f = new FolderBrowserDialog();
-
+        //Extensão do Arquivo Eletrônico.
         public String Extensao;
+        //Caminho completo até o arquivo (digitado pelo user).
         public String Caminho;
-        public Boolean Existe;
-
 
         public Arquivo(String Extensao, String CaminhoArquivo)
         {
             this.Extensao = Extensao;
             this.Caminho = CaminhoArquivo;
-            this.Existe = File.Exists(CaminhoArquivo);
         }
 
 
         public void ApagarLCK()
         {
-            String nomeArquivo = this.Caminho.Substring(0, this.Caminho.Length - 3) + "lck";
+            //Como é um arquivo LCK, esta linha abaixo troca a extensão do arquivo digitado.
+            this.Caminho = this.Caminho.Substring(0, this.Caminho.Length - 3) + "lck";
 
-            if (File.Exists(nomeArquivo))
+            //Se o arquivo existir.
+            if (File.Exists(this.Caminho))
             {
                 try
                 {
-                    File.Delete(nomeArquivo);
+                    //Deleta o arquivo
+                    File.Delete(this.Caminho);
                     MessageBox.Show("LCK apagado! Pague um café para o Ádamo!", "Apagar LCK");
                 }
                 catch (Exception e)
@@ -52,11 +53,22 @@ namespace Comos_Utilities.classes
 
         public void CopiarCRP()
         {
-            if (this.f.ShowDialog() == DialogResult.OK)
+            //Se o arquivo existir.
+            if (File.Exists(this.Caminho))
             {
-                String destFile = System.IO.Path.Combine(this.f.SelectedPath, Path.GetFileName(this.Caminho));
-                File.Copy(this.Caminho, destFile);
-                MessageBox.Show("Arquivo Copiado! Pague um café para o Ádamo!", "Copiar CRp");
+                //Abre a janela para o usuário selecionar um diretório e verifica se ele clicou em OK.
+                if (this.f.ShowDialog() == DialogResult.OK)
+                {
+                    //Faz uma concatenação entre o diretório que o usuário selecionou e o nome do arquivo que ele vai copiar.
+                    String destFile = System.IO.Path.Combine(this.f.SelectedPath, Path.GetFileName(this.Caminho));
+                    //Copia o arquivo.
+                    File.Copy(this.Caminho, destFile);
+                    MessageBox.Show("Arquivo Copiado! Pague um café para o Ádamo!", "Copiar CRp");
+                }
+            }
+            else
+            {
+                MessageBox.Show("Arquivo CRp não encontrado!", "Erro");
             }
         }
     }
